+++ src/components/blog/index.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { StyledContainer, StyledTitle, StyledText } from "./Blog.styled";
 import Blogpostcard from "@/components/common/blogpostcard";
 import { useState } from "react";
 import { categories } from "@/data/category";
@@ -14,10 +13,7 @@ function BlogPage({ posts }: any) {
 	const filteredPosts = select === "" ? posts : posts.filter((post: any) => post.category.trim() === select);
 
 	return (
-		<StyledContainer>
-			<StyledTitle> Blog </StyledTitle>
-			<StyledText style={{ textAlign: "center" }}>공유하고 싶거나 다시 보고 싶은 기술들을 정리합니다.</StyledText>
-
+		<>
 			<TagList categories={categories} select={select} onSelect={handleSelect} />
 
 			<div style={{ marginTop: "40px" }}>
@@ -25,7 +21,7 @@ function BlogPage({ posts }: any) {
 					<Blogpostcard date={post.date} title={post.title} des={post.description} slug={post._raw.flattenedPath} key={post._id} />
 				))}
 			</div>
-		</StyledContainer>
+		</>
 	);
 }
